+++ src/main/java/codesquad/codestagram/domain/article/exception/ArticleNotFoundException.java
@@ -1,4 +1,4 @@
-package codesquad.codestagram.domain.article;
+package codesquad.codestagram.domain.article.exception;
 
 public class ArticleNotFoundException extends RuntimeException {
     public ArticleNotFoundException(String message) {
