+++ src/main/java/codesquad/codestagram/domain/auth/exception/UnauthorizedException.java
@@ -1,4 +1,4 @@
-package codesquad.codestagram.domain.auth;
+package codesquad.codestagram.domain.auth.exception;
 
 public class UnauthorizedException extends RuntimeException {
     public UnauthorizedException(String message) {
