+++ src/main/java/codesquad/codestagram/domain/user/UserService.java
@@ -1,6 +1,8 @@
 package codesquad.codestagram.domain.user;
 
-import codesquad.codestagram.domain.auth.UnauthorizedException;
+import codesquad.codestagram.domain.auth.exception.UnauthorizedException;
+import codesquad.codestagram.domain.user.exception.DuplicatedUserException;
+import codesquad.codestagram.domain.user.exception.UserNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
