+++ src/main/java/codesquad/codestagram/domain/user/exception/DuplicatedUserException.java
@@ -1,4 +1,4 @@
-package codesquad.codestagram.domain.user;
+package codesquad.codestagram.domain.user.exception;
 
 public class DuplicatedUserException extends RuntimeException {
     public DuplicatedUserException(String message) {
